Changed name of generalTable to general-table, moved Empty
