+++ src/Routes/ImageManager/ImageTable.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import GeneralTable from '../../components/generalTable/GeneralTable';
+import GeneralTable from '../../components/general-table/GeneralTable';
 import PropTypes from 'prop-types';
 import { shallowEqual, useSelector } from 'react-redux';
 import { routes as paths } from '../../../package.json';
